fix(imu): 适配华轩阳 ICM-42605-HXY 寄存器映射并打通六轴数据闭环
- 根据官方手册更新寄存器协议，修正芯片 ID 寄存器地址为 0x01 (识别码 0x6A)
- 修正整机软复位逻辑 (0x4A 写入 0xA5) 与核心电源管理唤醒 (0x7D 写入 0x0E 开启三轴起振)
- 重构六轴数据突发读取起始地址为 0x0C，彻底解决因误读 0x1F FIFO 状态位导致的 4.00g 常数假死问题
- 优化 100Hz 高频主控节拍与 5Hz 串口日志等宽对齐排版
- 归档官方数据手册至文档/资料/ICM.pdf

+++ firmware_chassis/Inc/log.h
@@ -13,16 +13,16 @@
 #define LOG_CLR_BLUE "\033[34m"
 
 // 企业级格式化日志输出: [时间戳ms][级别][模块标签] 消息
-#define LOG_I(tag, fmt, ...)                                                                \
-    printf(LOG_CLR_GREEN "[%8u][I][%s] " fmt LOG_CLR_RESET "\r\n", bsp_get_ticks_ms(), tag, \
+#define LOG_I(tag, fmt, ...)                                                                                 \
+    printf(LOG_CLR_GREEN "[%8lu][I][%s] " fmt LOG_CLR_RESET "\r\n", (unsigned long)bsp_get_ticks_ms(), tag, \
            ##__VA_ARGS__)
 
-#define LOG_W(tag, fmt, ...)                                                                 \
-    printf(LOG_CLR_YELLOW "[%8u][W][%s] " fmt LOG_CLR_RESET "\r\n", bsp_get_ticks_ms(), tag, \
+#define LOG_W(tag, fmt, ...)                                                                                  \
+    printf(LOG_CLR_YELLOW "[%8lu][W][%s] " fmt LOG_CLR_RESET "\r\n", (unsigned long)bsp_get_ticks_ms(), tag, \
            ##__VA_ARGS__)
 
-#define LOG_E(tag, fmt, ...)                                                              \
-    printf(LOG_CLR_RED "[%8u][E][%s] " fmt LOG_CLR_RESET "\r\n", bsp_get_ticks_ms(), tag, \
+#define LOG_E(tag, fmt, ...)                                                                               \
+    printf(LOG_CLR_RED "[%8lu][E][%s] " fmt LOG_CLR_RESET "\r\n", (unsigned long)bsp_get_ticks_ms(), tag, \
            ##__VA_ARGS__)
 
 // 专用于上位机 (如 VOFA+ 的 FireWater 引擎) 绘图的纯净波形宏 (CSV 格式)
